+++ torii_boards/lattice/machxo2/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .tinyfpga_ax1 import TinyFPGAAX1Platform
+from .tinyfpga_ax2 import TinyFPGAAX2Platform
+
+__all__ = (
+	'TinyFPGAAX1Platform',
+	'TinyFPGAAX2Platform',
+)