+++ torii_boards/lattice/machxo3/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+from .machxo3_sk import MachXO3SKPlatform
+
+__all__ = (
+	'MachXO3SKPlatform',
+)